refactor(graph): extract MutationMixin into mutation.py (PR 3)
